cachegwfs: use redirect path only for stub file open
Change arguments passed to cachegwfs: <source> <target> <redirect>

'source' is the real path where lookup and file handles are done.
'target' is the mount point.
'redirect' is the path where open of stubs is redirected to.

This changes the root inode fd to point at the source root dir
instead of the redirect dir and simplifies the code.