cachegwfs: safe config reload
Setting string values in redirect struct is not async signal safe.
This may end up with not redirecting stubs or even with crash.

Change to lazy read of config file on first access to Redirect object.
Singal handler just sets a flag to reload config on next from thread
context.