cachegwfs: reply error instead of abort()
Handle internal errors more gracefully.

On many requests, we need to resolve ino to open fd.
That can fail for 2 reasons:
1. inode not found in inodes map - reply ENOENT
2. cannot open fd by file handle - reply ESTALE

All operations start with:
    InodeRef inode(get_inode(ino));

On first step, instead of abort() on lookup in get_inode(ino),
return a dead inode object (i.e. fs.nullinode).

On second step, instead of abort() on open_by_ino() in InodeRef(),
set an internal error in the InodeRef object.

Finally, all operations check inode.error() to send a reply and
return if an internal error was set. If the InodeRef object references
a dead inode ENOENT is returned.