cachegwfs: use shared_ptr in inodes map
It is not safe to access inodes in the map not under fs.mutex lock
because entry may be erased at any time.

Convert the inodes entries to shared_ptr and get rid of the unscoped
inode mutex unlock before erase from map.

Converting nlookup to atomic counter is left for another time.