cachegwfs: disable writeback cache by default
Currently all caching is bundled together with a single option
"--nocache" and by default all caches are enabled with 1s timeout.

writeback cache has more implications when enforcing out of quota
and for write throttling in server, so decouple writeback cache
from the global nocache option and disable it by default.
It can be enabled with option "--wbcache" as long as "--nocache"
is not provided to override it.