cachegwfs: always use magic procfs path for *xattr ops
cachegwfs code depends on an out-of-tree kernel patch to allow
f*xattr ops on an O_PATH fd.

The author of the patch has withdrawn the proposal [1], because
there is another way to perform these ops without a kernel change.

Always perform *xattr ops on the procfs magic link path instead of
using f*xattr on the O_PATH fd.

[1] https://lore.kernel.org/linux-fsdevel/CAOssrKeV7g0wPg4ozspG4R7a+5qARqWdG+GxWtXB-MCfbVM=9A@mail.gmail.com/