cachegwfs: make sure not to reuse inode object for reused inode
An XFS inode number does not uniquely identify an object.
The inode generation needs to match as well as inode number.

Ideally, cachegwfs would get a FORGET message on the old inode
before the inode numebr can be reused by XFS, but there is no
guaranty that will happen. For example, inode can be unlinked
on XFS underneath cachegwfs.

Store the inode generation in the Inode object and make sure that
xfs inode generation agrees with the Inode object when Inode object
is found by inode number in lookup.

We always open the file using the real file handle found on lookup
instead of by inode number alone, which may open a new file with the
reused inode number.