cachegwfs: use FSBULKSTAT ioctl to get generation from inode number
This is to be used on unmodified kernel instead of the XFS patch
to decode fake file handle.

bulkstat only works on xfs directly and not on snapshot fs and we do
not need to use it if we have a kernel with patched XFS.

If we fail to bulkstat (because fs is snapshot fs), we try to open
by fake file handle with zero generation, If that succeeds we disable
future attempts of bulkstat.