cachegwfs: create objects with redirected path
as_user() assumes that the first attempt to create the object has
populated the parent dir, so it can retry to create the object in
bypass dir, but the permission check is done before calling FUSE,
so directory can still be a stub at this point.

If the original create operation was executed in redirect path,
perform the retry+chown also in redirect path.