cachegwfs: abstract get_file_fd()
fi->fh is a different object for open files and open directories.
do_setattr() assumes that fi is an open file (and not directory) when
calling get_file_fd(fi) for ftruncate, fchmod and futimens.

This happens to be true, because fuse kernel driver only ever sends a
setattr request with an open file for ftruncate of a regular file,
but that code is not pretty, so fix it.