cachegwfs: relax error messages for possible use cases
It is not a fatal error if we detect node number reuse or if file
cannot be opened due to flock.

The application will get an error, but these things do happen so
no need for an alarming error/warning log which causes libfuse tests
to fail.

Add some debug info about inode's generation