cachegwfs: use open_rw op for create operation
create() is a private case of open_rw operation.
We need to use the same op as we use in check_safe_fd()
after file is opened in create() operation, so use open_rw.

This changes the debug log from "create(foo): ..." to
"open_rw(foo): ...", but the log is still different from the
debug log for open of an existing file - "open_rw(): ..."), because
dirfd of the create() operation is that of the parent directory and
dirfd of the open() operation is that of the opened file itself.