notifyfs: record parent in index state
An indexed state is not friendly to hardlinks, because the state
is stored in an inode, but different hardlinks may have different
"indexed path" state.

Therefore, hardlinks are handled as best effort - the inode indexed
state reflects the indexed state of the last path by which the inode
was looked up.

Rename does not change the parent recorded in the inode state.
The fact that an inode itself is indexed or new does not change on
rename and both the old and new parent paths must be indexed before
rename, so the indexed state itself should not be changed on rename.

The next lookup after rename will find the inode with the new parent
and reset the indexed state with the new parent.

If file was renamed in source not via the passthrough fs, lookup may
find an inode with inconsistent state.  In this case, the state will
be reset to prevent modification to the inode until the next lookup().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>