cachegwfs: error on failure to redirect
When operation needs to be redirected and redirect fails because
it was not able to resolve a valid realpath, fail the operation
instead of executing it on the real fd.

This may happen for an fd open_by_ino() with disconnected path.

Until commit 40e3c61 ("cachegwfs: check stub after flock") this
situation could result in reading the zero content of a stub file.
Since that commit, this situation wrongly reports the message:
  INFO: file open raced with evict.

This is because do_open() fails to redirect and concludes that the
file is not a stub, but after opening the real file it discovers
that it is in fact a stub.