cachegwfs: store a connecatable file handle for non-dir inode
File handles returned by name_to_handle_at() are "non-connectable"
meaning that opening an fd with open_by_handle_at() may return an fd
with unknown parent and unknown path.

An fd with unknown path cannot be redirected and if it needs to be
redirected, an error (ENOENT) will be returned for the attempted
operation.

To avoid that, if kernel has patch to support opening "connectable"
file handles, then we store "connectable" file handle that encodes
also the parent <ino;gen>.

When open by a "connectable" file handle succeeds, it is expected
that the fd path is known. If the file was moved to another parent
via cachegwfs, we call lookup to store a "connectable" file handle
with the new parent.

If the file was moved to another parent not via cachegwfs, open by
an old "connectable" file handle may return an fd with the new path
or an fd with an unknown path that we can use as long as it does not
need to be redirected.