cachegwfs: avoid unintended follow of redirected symlinks
Some syscalls cannot operate on O_PATH fd so we use use magic /proc
symlink to reference inode and then we cannot use NOFOLLOW.

When the operation subject is a symlink inode, the magic /proc symlink
is followed to the symlink inode and the symlink inode is not followed.
However, if the operation is redirected, the "follow" semantics may
lead to unintended following of the redirected symlink.

To avoid that, do not redirect opertions on symlink inodes unless
the syscalls are supposed to follow symlinks (e.g. truncate).