cachegwfs: add support for sub-folder id xattrs
In addition to folder ids that are determined by name of first level
subdir, support also setting a sub-folder id on an inner directory by
setting a sub-folder id value in xattr.

The sub-folder id in xattr feature is enabled by setting the value of
redirect_filder_id_xattr to the name of the xattr to read the folder id.

This option replaces the xfs projid sub-folder ids.