cachegwfs: keep open O_PATH fds for all inodes by default
cachegwfs stores connectable file handles for all inodes in cache, so
it can open an O_PATH fd on demand, but if memory usage is not
constrained, it is more efficient to keep the open O_PATH fds, because
it helps to keep the underlying filesystem inodes in cache as well.

The command line argument '--nokeepfd' can be used to disable this
behavior and keep open O_PATH fds only for first level directories and
in case we have no connectable file handles.