cachegwfs: store version along with folder id
cachegwfs stores a folder id in all inodes, which is inherited from
parents on lookup.

Read folder_id_version from config and store it in inodes along with the
inherited folder id, so that users can invalidate the inherited folder
ids of all children when they change a folder id of some parent dir.