cachegwfs: revalidate folder_id on open for write
On open for write, if open was not redirected, we check if the inherited
folder id version of the inode is uptodate w.r.t the config version.

If it is not uptodate, we open the redirect fd and query the uptodate
folder id from the redirected fd and update the inode.

Then we check if the uptodate folder id needs to redirect writes, and
if it does, we use the redirected fd for the open.

With non-empty folder_id_xattr, folder_id_version > 0 is expected.
folder_id_version == 0 is a special configuration for testing
"always query uptodate folder id" on open for write.

On lookup, we query folder_id from xattr only if parent's folder id
is not uptodate and folder_id_version != 0.