cachegwfs: fix reconnect of inodes with disconnected path fds
The "keepfd" feature has a bug - when a non-dir inode is found with
unknown path (from NFS) the disconnected path fd is kept as long lived fd
and not updated to connected path fd on later lookup of inode by name.

Similarly, the inode with unknown path has an unknown folder id which is
also not updated on later lookup of inode by name as it should.