cachegwfs: propagate flush to redirect fd
We changed the meaning of FOPEN_NOFLUSH flag to
"do not flush on close if there was no io on fd".

If we get a flush() callback on a rdonly fd, it means there
was io (maybe kernel passthrough io) on this fd.

In that case, we want to propagate flush() on close of the the
rdonly redirect fd, so issue a zero legth read on the redirect
fd to clear the FOPEN_NOFLUSH flag on that fd.