Close passthrough backing files on fuse file release
Test reference counted passthrough backing files.
A backing file needs to be mapped before open reply and unmapped
sometime after the client receives the open response.

We cannot unmap the backing file right after sending the open reply,
because the open reply with the backing id sits in the request queue
until the client handles the response and by that time, the mapped
backing file id will have been gone.

To make things simpler, we stash the backing file id in the passthrough
filesystem fh and unmap it on release.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>