cachegwfs: fix incorrect lookup count after rename
cachegwfs added an internal call to do_lookup() for internal reasons
but the kernel does not know about this lookup, so the nlookup count
of the cachegwfs inode diverges from the kernel's nlookup count.

When kernel sends a FORGET message, it carries a too low lookup count
and the cachegwfs inode is not destoyed.

decrement the lookup cachegwfs inode lookup count after the internal
do_lookup() in rename() to fix this missmatch.