Enable passthrough of multiple fuse files to the same backing file
When setting up an auot close backing file, the kernel will attach the
backing file to the fuse file and close it on release of the fuse file.

When setting up an shared backing file id, the server is resposible to
to close it explicitly by calling fuse_passthrough_close() when the
shared backing file is no longer needed.

Modify the passthough_hp example to use a shared backing file for all
read-only files of the same inode and unique backing files for writeable
files.

The server closes the inode shared backing file on the last file close of
an inode, so even if there are only writeable files open for the inode,
the shared backing file remains attached to the inode and reused by new
read-only opens.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>