cachegwfs: simplify logic of redirect by stub xattr
For some reason, the implementation of should_redirect_fd() returns false
if an operation is not in the redirected set, so the config options
redirect_{read,write}_xattr need to set redirect_op = open_{ro,rw}
just to get to the stub xattr check.

This makes the behavior of setting both redirect_op = open_rw rule and
redirect_write_xattr quite odd when it will not redirect if stub xattr
is not found despite the explicit redirect_op = open_rw rule, but we
never set both those options, we only even set one of them.

The logic becomes much simpler if should_redirect_fd() does the opposite:
return true if an operation is in the redirected set, then check for the
stub xattr for selected operations and return true if it is found.

A similar simplification is done for the redirect_read_once (AV)
config option: instead of implying redirect_op = copy, check also
redirect_once_enabled() to decide if redirect fd should be opened.