cachegwfs: add config redirect_writedir_xattr
This is a variant of redirect_write_xattr for directories in a
similar way that redirect_readdir_xattr is redirect_read_xattr for
directories.

The meaning of "writedir" is any operation that changes the directory
content, so: mkdir, rmdir, mknod, rename, etc.

We do not need to worry about checking stub xattr on the directory for
those operations, because they follow a lookup operation where already
checks for stub xattr directory and redirect access() to cafs.