cachegwfs: redirect create operation with redirect_writedir_xattr
create() operation is actually an atomic create+open.
instead of checking stub xattr on both parent directory and created
file, change the code to check the stub redirect_writedir_xattr on
the parent directory in do_create() and the stub redirect_write_xattr
on the created file in fd_open().