passthrough: store a connecatable file handle for non-dir inode
File handles returned by name_to_handle_at() are "non-connectable"
meaning that opening an fd with open_by_handle_at() may return an fd
with unknown parent and unknown path, that will prevent passthrough
modules from working with paths.

If kernel supports encoding "connectable" file handles, we store
"connectable" file handle that encodes also the parent <ino;gen>.

When open by a "connectable" file handle succeeds, it is expected
that the fd path is known. If the file was moved to another parent
via fuse, we call lookup to store a "connectable" file handle
with the new parent.

If the file was moved to another parent not via fuse, open by an old
"connectable" file handle may return an fd with the new path or an fd
with an unknown path that modules can use as long as they do not need
to know the path.