passthrough: allow each module to store/fetch state in file
The state is a shared_ptr<void> state and the state object destructor is
assigned by the module that allocated the shared_ptr object.

Unlike the pre-module inode state, the file state has no multithreaded
protection in the library - it is the module's responsibility to protect
its own state from concurrent modification, for example, by setting state
at file open time and not modify it afterwards.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>