passthrough: generalize fuse_passthrough_module registration
Generalize the registration helpers so user can chain multiple modules.

Unlike the high level fuse_module, we do not chain the passthrough
modules in a linked list.  Instead, we chain the implemented operations
so modules do not need to implement all the operations.

The "default" passthough fs registers itself as the head of the module
chain.  It's next member is the head of the chain and its oper member
is the tail of the chain containing the default fs operations.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>