cachegwfs: create passthrough module
Parse command line arguments and config file and call
fuse_passthrough_main().

cachegwfs is intended to be used by hierarchic storage manager (HSM)
to redirect some of the operations on some of the files to an alternative
tiered storage location.

The command line option --redirect can be used to test redirect all
the operations.

No operations are currently intercepted, so redirect is not yet
implemented.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>