cachegwfs: add config redirect_read_once_{older,grace}
This will cause redirecting a file on open if the file was last accessed
more than $grace seconds before $older timestamp or if file was modified
after it was last accessed.

If 0 or no $older timestamp is provided, open will be redirected if file
was last accessed more than $grace seconds ago.

$grace may be omitted or set to 0 for no grace period, but normally it
should be set to 86400 (1 day) to align with the default relatime
policy of updating atime.

This guarantees that at least one read per file is redirected after
setting a new $older timestamp and the first read from a file after
it is created or modified is also redirected.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>