cachegwfs: add config redirect_writedir_xattr
This is a variant of redirect_write_xattr for directories in a
similar way that redirect_readdir_xattr is redirect_read_xattr for
directories.

The meaning of "writedir" is any operation that changes the directory
content, so: mkdir, rmdir, mknod, rename, etc.

We do not need to worry about checking redirect_readdir_xattr on the
directory for those operations, because they follow a lookup operation
where already checks for the redirect_readdir_xattr on the directory.

create() operation is actually an atomic create+open which is going to
change both the parent directory and the opened file.
Instead of checking xattr on both parent directory and created file,
check the redirect_writedir_xattr on the parent before create() and
check the redirect_write_xattr on the open file after create().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>