passthrough: allow each module to store/fetch state in inode
The state is a 64bit integer that can be used by modules to store an
arbitrary value or a refernce an allocated state object.

Modules that store pointers to allocated state objects need to implement
the forget() operation to free the state object before inode is freed.

An optional filler() callback can be used to init or update the state
in the context of inode lock with a reference to an inode with O_PATH fd.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>