passthrough: allow each module to store/fetch state in inode
The state is a shared_ptr<void> inode state and the state object
destructor is assigned by the module that allocated the shared_ptr
object.

An optional filler() callback can be used to init or update the state
in the context of inode lock with a reference to an inode with O_PATH fd.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>