passthrough: allow opt-out for read/write passthrough per file
The passthrough library handles methods that are not implemented by
modules without calling the modules, but when modules implement methods
the passthrough library will call those methods.

The default behavior is changed to not call the read/write/readdir
methods of modules unless they explicitly opt-out of library passthrough
per open file.

A modules that wants its read/write/readdir methods to be called, needs
to clear the fi->passthrough_{read,write} flags at open() time.

This method is opposite to the kernel passthrough API, where the default
is that read/write are not passthrough by kernel and the server needs to
opt-in for kernel read/write passthrough per open file.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>