cachegwfs: take shared flock on non-redirected file open
cachegwfs redirects opens of stub files with evicted content.
A shared flock is taken on open of non-stub files to prevent eviction
of content while file is open.

There is a TOCTOU race in cachegwfs that can lead to:
- Check if file is stub => no redirect
-- File is locked and eviceted by Evictor
- File is locked by cachegwfs (but is already a stub)
- Stub is read/written to directly by cachegwfs

Another possible race:
- New file is being created and opened via cachegwfs
-- Stub of same name is populated
- do_create() finds the existing stub, opens and locks it (no redirect)
- Stub is read/written to directly by cachegwfs

To fix this race, when file is opened without redirect, check
that it is not a stub after taking the lock and return EAGAIN if
a race was detected.

To ensure consistency redirect also create() when configured to
redirect open_rw.  Note that when checking stub before create(),
dirfd of the create() operation is that of the parent directory and
fd result of create() operation is that of the opened file itself.
Nevertheless, create() should not result in an opened stub file.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>