cachegwfs: open redirect fd at open/create time
If we need to redirect copy_file_range, instead of opening redirect fd
during copy_file_range(), open the redirect fd at open()/create() time
and use it in copy_file_range().

This changes the meaning of redirect_op=copy - Instead of forcing the
use of redirect fds for the copy operation, it means that redirect fds
are always available and that we use the redirect fds for copy in case
one of the fds was redirected at open.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>