notifyfs: do not index directories newer than index
If filesytem supports birth time (e.g. xfs, ext4), check btime and
avoid indexing directories that were created after the index directory.

With this change, the test test_passthrough_hp with notifyfs does
not index any directory when run on a filesystem that supports birth
time, because all the directories created in the test are newer than
the index directory.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>