notifyfs: mark moved directories in index
On rename, in addition to indexing the src parent path
and target parent path, if the moved inode is a directory,
we index the moved directory inode and mark it "moved" by
setting an xattr on the index entry.

We can only do that when FUSE ino (nodeid) is the same as
the source st_ino. Otherwise, we return EXDEV to let userspace
fallbak to recursive move.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>