passthrough: compose and reconnect a connectable file handle
When requested to provide fds with connected path,
If kernel does not support encoding "connectable" file handles, we use
the encoder to compose a "connectable" file handle from the parent and
child file handles.

When opening an O_PATH fd from the file handle, if the fd path is
unknown, try to get an fd with known path by iterating the parent dir
to look for the child entry by name.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>