passthrough: use kernel passthrough mode for read/write operations
When passthrough module delegates read/write to passthrough library,
the library may delegate read/write to the kernel if kernel passthrough
feature is supported.

If kernel passthrough is supported and requested once on an inode, all
further opens will use the same backing file.  The passthrough module
can still opt-out of kernel passthrough for a specific fd open, but
in that case, passthrough library will open the file in dio mode.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>