notifyfs: index directories before modification
Modification is still not allowed unless all ancestors are also indexed.

With this change, the only test case in test_passthrough_hp that fails
with notifyfs is tst_open_write(), because opening a file for write
requires to index the parent directory first.

All the rest of the tests, like test_create() are passing, because
notifyfs is called with root directory inode + filename, so it is able
index the root directory.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>