cachegwfs: fix double fd close on release
cgwfs_release() is calling the library release method to close the main fd
and close_file_redirect_fd() to close the optional redirect fd.
If the redirect fd is also the main fd, do not close them twice.

There is never a need to redirect close, so remove the OP_RELEASE
redirect.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>