notifyfs: fix implementation of index state reset
The implementation of reset() with set(IDX_INIT) was incorrect
because it is using fetch_or() which does not clear the state bits.

Create a new method reset() with correct implementation, make set()
method void check IDX_VALID() in test() and change the name of the
method bits() to get().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>