notifyfs: test changing index dir
Test that index entries are created in initial index and that
after change of index dir, index entries are created only
in the new index.

We change the index before test_syscalls which creates some subdirs.
Those subdirs are new so only recoded in index with --index_all option
so add this option to the nocache run, which is the only test variant
that runs test_syscalls.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>