notifyfs: move initialization of root inode idx state
The root inode index state is initialized with the IDX_PARENT
state because unlike other inodes, it is not looked up from
another parent dir.

We need to make sure that this flag is set every time the root inode
index state is checked and not only when the state is allocated, because
when we are changing an index, the root inode index state is already
allocated, but its state w.r.t the new index needs to be re-initialized.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>