notifyfs: fix wrong call to dir_is_new()
In the call index_move_path_at(), wehn a subdir at dirfd+path is being
moved, dir_is_new() was supposed to check if the moved subdir btime is
newer than index dir btime, but instead it was checking the btime of its
parent.

Pass an optional path argument to dir_is_new() and instead of passing
the ino argument (for printing) get it back from get_dir_btime_nsec(),
which anyway needs to call statx() for btime.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>