notifyfs: prepare for replacing the index dir
Instead of keeping index_path and index_btime in Notifyfs object,
keep them in a separate Index object which is defered from a
shared_ptr.

All notify operations take a reference to the Index shared_ptr and
perform the entire scope with this Index pointer.

This will allow to replace the Index object of Notifyfs to an Index
with a different path without impacting operations in progress.

Canonicalize the provided index path with realpath(), so that we can
skip changing Index object for the same effective path.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>