notifyfs: invalidate index state on index dir change
The index state bits cache whether a directory and its parents
are indexed in a particular index dir.

When changing the index dir those cached states should be invalidated.

Store the index id along with the index state bits, and test them
w.r.t the current index id.  When index id is changed, it will auto
invalidate the old index state.

NOTE: two threads could be operating on the same directory
concurrently, one still holding a reference to the old index and
the other holding a reference to the new index, so they could be
thrashing each others index state cache, but only over a relatively
short window of time. Next change will deal with that.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>