notifyfs: record change in index on close of rw fd
Record a change on close of rw fd in curr index.
This is needed in case fd was opened in a time of a prev index
and written and closed at time of a new index.

This does not cover the case that file is written at time of new
index but not closed.

Recording the change is both best effort and async, because fuse
close is non blocking.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>