notifyfs: implement xattr "user.notifyfs.old_rwfiles"
getxattr of this xattr without a buffer returns <number of rwfile
in the old list> * sizeof(ino_t).

getxattr with a buffer, fills the buffer with inode numbers of files
from old rwfiles list while moving them over to current list.

This is a method for userspace to iterate all the files that have been
opened rw during previous index and need to be recored in new index.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>