cachegwfs: treat failure to redirect as fatal error
When cachegwfs decides to redirect an operation to the redirect path,
it must know the path of the target.  In case of disconnected path
redirect will fail.  In that case, cachegwfs used to emit an error and
perform the operation on the non-redirected path.

This is risky, because applications depend on redirecting operations
for correctness and safety.

Upon failure to compose a redirect path, return a fuse_path_at object
with invalid path, so an attempt to use that path in a syscall will fail.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>