readdir_inode: support reading a list of directories from stdin
mbits/khash determined from the first directory.
each directory read updates bloom filter and prints the total
number of entries and number of collisions.

This can be used to simulate false positive on a huge directory.
Use directory inode number as seed to not treat name collisions across
different directories as a match.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>