cachegwfs: open the redirect fd with original user credentials
When opening the redirect fd in cafs path, cafs may deny the open
due to archive folder restrictions, but it needs to know the original
uid/gid ir order to make the correct decision.

The same is true for the getxattr issued on this fd to check if read
and write should be using the redirect fd (a.k.a. NoBypassRW).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>